requirements.txt
# GlitchRealityAI — Python dependencies
# All packages listed here are free to use.

# Core HTTP
requests>=2.31.0
httpx>=0.27.0

# Config parsing
PyYAML>=6.0.1
pydantic>=2.7.0

# Free TTS (Microsoft Edge TTS — no API key needed)
edge-tts>=6.1.12

# Async support
anyio>=4.4.0

# Hugging Face (for InferenceClient where needed)
huggingface-hub>=0.23.4
gradio-client>=1.0.0

# Prompt management / utilities
python-dateutil>=2.9.0

# Optional: OpenAI TTS upgrade path
# openai>=1.35.0

# Optional: ElevenLabs upgrade path
# elevenlabs>=1.1.0

# Dev / testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
